requirements.txt
pandas==1.4.2
requests==2.27.1
websocket_client==1.3.2
